Wiring the beans using direct method call between the @Bean methods
The @Bean method which is invoked in other @Bean methods, e.g. parrot()
is only called once. Spring is smart enough to understand that by calling
the parrot() method, you want to refer to the parrot bean in its context.
When we use the @Bean annotation to define beans into the Spring context,
Spring controls hw the methods are called and can apply logic above the
method call: If the parrot bean already exits in the context, then instead
of calling the parrot() method, Spring will directly take the instance
from its context. If the parrot bean does not exist in the context,
Spring calls the parrot() method and returns the bean.