[4.3] Focusing on object responsibilities with stereotype annotations
Spring offers us the @Service annotation to mark a component that takes
the responsibility of a service and the @Repository annotation to mark
a component that implements a repository responsibility. All three
@Component, @Service, and @Repository are stereotype annotations and
instruct Spring to create and add an instance of the annotated class
to its context.

The purpose of different annotations are to provide better visibility
for any developer reading the class.