CH7: A simple web app with @Controller
@Controller method provides the view name and data that needs to be rendered in view.
The dispatcher servlet finds the view (by default, in resources/static) and renders
it with the data provided by the @Controller. To find the view to render, the dispatcher
servlet uses a component named view resolver.

The view can be any files, not just html. Displaying the file is handled by client,
a.k.a. Web Browser